requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
-e git+https://github.com/serengil/deepface.git@774012ab88ba588e68fee3ccf04e9923d3befd67#egg=deepface
distlib==0.3.9
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.55.3
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
mypy-extensions==1.0.0
namex==0.0.9
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
onnxruntime==1.21.1
onnxruntime-gpu==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.4
pycodestyle==2.12.1
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
sklearn-compat==0.1.3
soupsieve==2.6
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.5.0
tifffile==2025.3.30
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
virtualenv==20.28.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yellowbrick==1.5
